+++ src/main/java/com/github/halvra/opencell/actions/SelectEnvironmentComboBoxAction.java
@@ -7,25 +7,19 @@ import com.github.halvra.opencell.settings.model.Environment;
 import com.intellij.icons.AllIcons;
 import com.intellij.ide.DataManager;
 import com.intellij.openapi.actionSystem.*;
-import com.intellij.openapi.actionSystem.ex.ComboBoxAction;
 import com.intellij.openapi.options.ShowSettingsUtil;
-import com.intellij.openapi.project.DumbAware;
 import com.intellij.openapi.project.Project;
 import com.intellij.ui.SizedIcon;
-import com.intellij.ui.components.panels.NonOpaquePanel;
 import com.intellij.ui.scale.JBUIScale;
-import com.intellij.util.ui.JBUI;
-import com.intellij.util.ui.UIUtil;
 import icons.Icons;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import javax.swing.*;
-import javax.swing.border.Border;
 import java.awt.*;
 import java.awt.event.ActionEvent;
 
-public class SelectEnvironmentComboBoxAction extends ComboBoxAction implements DumbAware {
+public class SelectEnvironmentComboBoxAction extends EnvironmentComboBoxAction {
     private static final String BUTTON_MODE = "ButtonMode";
 
     public static final Icon CHECKED_ICON = JBUIScale.scaleIcon(new SizedIcon(AllIcons.Actions.Checked, 16, 16));
@@ -57,11 +51,6 @@ public class SelectEnvironmentComboBoxAction extends ComboBoxAction implements D
             presentation.putClientProperty(BUTTON_MODE, Boolean.TRUE);
             presentation.setText(OpencellBundle.message("toolbar.environment.add"));
         }
-    }
-
-    @Override
-    protected boolean shouldShowDisabledActions() {
-        return true;
     }
 
     @NotNull
@@ -101,36 +90,19 @@ public class SelectEnvironmentComboBoxAction extends ComboBoxAction implements D
                 return !Boolean.TRUE.equals(presentation.getClientProperty(BUTTON_MODE));
             }
         };
-        NonOpaquePanel panel = new NonOpaquePanel(new BorderLayout());
-        Border border = UIUtil.isUnderDefaultMacTheme() ?
-                JBUI.Borders.empty(0, 2) : JBUI.Borders.empty(0, 5, 0, 4);
-
-        panel.setBorder(border);
-        panel.add(myButton);
-        return panel;
+        return createCustomComponent(myButton);
     }
 
     @Override
-    protected @NotNull DefaultActionGroup createPopupActionGroup(JComponent button) {
-        final DefaultActionGroup allActionsGroup = new DefaultActionGroup();
-        final Project project = CommonDataKeys.PROJECT.getData(DataManager.getInstance().getDataContext(button));
-        if (project == null) {
-            return allActionsGroup;
-        }
-
-        final ProjectSettingsState settings = ProjectSettingsState.getInstance(project);
-
-        allActionsGroup.addSeparator();
-        settings.getEnvironments().forEach(environment -> allActionsGroup.add(new SelectEnvironmentAction(project, environment, environment.isPreferred())));
-
-        return allActionsGroup;
+    protected AnAction selectAction(Project project, Environment environment) {
+        return new SelectEnvironmentAction(project, environment);
     }
 
     private static final class SelectEnvironmentAction extends AnAction {
         private final Project project;
         private final Environment environment;
 
-        SelectEnvironmentAction(final Project project, final Environment environment, boolean selected) {
+        SelectEnvironmentAction(final Project project, final Environment environment) {
             this.project = project;
             this.environment = environment;
 
@@ -139,8 +111,8 @@ public class SelectEnvironmentComboBoxAction extends ComboBoxAction implements D
             presentation.setText(name, false);
             presentation.setDescription(OpencellBundle.message("select.0", name));
 
-            presentation.setIcon(selected ? CHECKED_ICON : Icons.OPENCELL);
-            presentation.setSelectedIcon(selected ? CHECKED_SELECTED_ICON : Icons.OPENCELL);
+            presentation.setIcon(environment.isPreferred() ? CHECKED_ICON : Icons.OPENCELL);
+            presentation.setSelectedIcon(environment.isPreferred() ? CHECKED_SELECTED_ICON : Icons.OPENCELL);
         }
 
         @Override