+++ src/main/java/com/github/halvra/opencell/services/OpencellApiService.java
@@ -1,0 +1,74 @@
+package com.github.halvra.opencell.services;
+
+import com.github.halvra.opencell.settings.model.Environment;
+import com.github.halvra.opencell.utils.HttpUtil;
+import com.google.gson.Gson;
+import com.intellij.openapi.diagnostic.Logger;
+import org.apache.http.HttpEntity;
+import org.apache.http.client.CredentialsProvider;
+import org.apache.http.client.ResponseHandler;
+import org.apache.http.client.methods.HttpPost;
+import org.apache.http.client.methods.HttpUriRequest;
+import org.apache.http.entity.ContentType;
+import org.apache.http.entity.StringEntity;
+import org.apache.http.impl.client.BasicCredentialsProvider;
+import org.apache.http.impl.client.CloseableHttpClient;
+import org.apache.http.impl.client.HttpClientBuilder;
+import org.apache.http.util.EntityUtils;
+import org.meveo.api.dto.ScriptInstanceDto;
+import org.meveo.api.dto.response.ScriptInstanceReponseDto;
+
+import java.io.IOException;
+
+public class OpencellApiService {
+    private static final Logger LOGGER = Logger.getInstance(OpencellApiService.class);
+
+    private final Gson gson;
+    private final Environment environment;
+
+    private OpencellApiService(Environment environment) {
+        this.gson = new Gson();
+        this.environment = environment;
+    }
+
+    public static OpencellApiService getInstance(Environment environment) {
+        return new OpencellApiService(environment);
+    }
+
+    public ScriptInstanceReponseDto createOrUpdateScript(ScriptInstanceDto scriptInstanceDto) {
+        return post("/api/rest/scriptInstance/createOrUpdate", scriptInstanceDto, ScriptInstanceReponseDto.class);
+    }
+
+    private <T> T post(String path, Object body, Class<T> returnType) {
+        HttpPost httpPost = new HttpPost(environment.getUrl() + path);
+        httpPost.setEntity(new StringEntity(gson.toJson(body), ContentType.APPLICATION_JSON));
+
+        return execute(httpPost, returnType);
+    }
+
+    private <T> T execute(HttpUriRequest request, Class<T> returnType) {
+        request.setHeader("Authorization", "Basic " + environment.getAuthorization());
+
+        try (CloseableHttpClient client = getClient()) {
+            ResponseHandler<T> responseHandler = response -> {
+                HttpEntity entity = response.getEntity();
+                return entity != null ? gson.fromJson(EntityUtils.toString(entity), returnType) : null;
+            };
+
+            return client.execute(request, responseHandler);
+        } catch (IOException e) {
+            LOGGER.error("Failed to deploy script to target environment", e);
+            return null;
+        }
+    }
+
+    private CloseableHttpClient getClient() {
+        final HttpClientBuilder clientBuilder = HttpClientBuilder.create();
+        final CredentialsProvider credentialsProvider = new BasicCredentialsProvider();
+
+        clientBuilder.setDefaultCredentialsProvider(credentialsProvider);
+        HttpUtil.configureProxy(environment.getUrl(), clientBuilder::setProxy, credentialsProvider::setCredentials);
+
+        return clientBuilder.build();
+    }
+}