+++ src/main/java/com/github/halvra/opencell/tasks/DeployScriptToEnvironmentTask.java
@@ -1,28 +1,19 @@
 package com.github.halvra.opencell.tasks;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.halvra.opencell.OpencellBundle;
-import com.github.halvra.opencell.dto.ScriptInstanceDto;
+import com.github.halvra.opencell.OpencellNotifier;
+import com.github.halvra.opencell.services.OpencellApiService;
 import com.github.halvra.opencell.settings.model.Environment;
-import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.progress.ProgressIndicator;
 import com.intellij.openapi.progress.Task;
 import com.intellij.openapi.project.Project;
-import org.apache.http.HttpStatus;
-import org.apache.http.client.methods.CloseableHttpResponse;
-import org.apache.http.client.methods.HttpPost;
-import org.apache.http.entity.ContentType;
-import org.apache.http.entity.StringEntity;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-
-import java.io.IOException;
+import org.meveo.api.dto.ActionStatusEnum;
+import org.meveo.api.dto.ScriptInstanceDto;
+import org.meveo.api.dto.response.ScriptInstanceReponseDto;
 
 public class DeployScriptToEnvironmentTask extends Task.Backgroundable {
-    private static final Logger LOGGER = Logger.getInstance(DeployScriptToEnvironmentTask.class);
-
     private final ScriptInstanceDto scriptInstance;
     private final Environment environment;
 
@@ -38,17 +29,16 @@ public class DeployScriptToEnvironmentTask extends Task.Backgroundable {
         indicator.setIndeterminate(true);
         indicator.setText(OpencellBundle.message("tasks.deployScript.progress", scriptInstance.getCode(), environment.getName()));
 
-        CloseableHttpClient client = HttpClients.createDefault();
-        HttpPost httpPost = new HttpPost(environment.getUrl() + "/api/rest/scriptInstance/createOrUpdate");
-        httpPost.setHeader("Authorization", "Basic " + environment.getAuthorization());
-        try (client) {
-            httpPost.setEntity(new StringEntity(new ObjectMapper().writeValueAsString(scriptInstance), ContentType.APPLICATION_JSON));
-            CloseableHttpResponse response = client.execute(httpPost);
-            if (response.getStatusLine().getStatusCode() != HttpStatus.SC_OK) {
-                LOGGER.warn("Failed to deploy script to target environment: " + response.getStatusLine().getStatusCode());
+        ScriptInstanceReponseDto response = OpencellApiService.getInstance(environment).createOrUpdateScript(scriptInstance);
+        if (response == null || response.getActionStatus().getStatus() != ActionStatusEnum.SUCCESS) {
+            String errorMessage;
+            if (response == null) {
+                errorMessage = OpencellBundle.message("tasks.deployScript.unknownError", environment.getName());
+            } else {
+                errorMessage = OpencellBundle.message("tasks.deployScript.error", environment.getName(), response.getActionStatus().getMessage());
             }
-        } catch (IOException e) {
-            LOGGER.error("Failed to deploy script to target environment", e);
+
+            OpencellNotifier.notifyError(myProject, errorMessage);
         }
     }
 }